fix: refactor out some config to /lsp/roslyn.lua
